feat: add daily one-word function and integrate ky HTTP library
The ky HTTP library was introduced into the project and the http tool module was created to handle HTTP requests uniformly. A new quote module was added to obtain daily word data and displayed in Home.vue. The daily word data will be cached for 1 hour to reduce unnecessary API calls.